Parse end positions in coq, nix, jsonnet, phpmd, proselint, rust, tslint
The second column returned by Coq is an offset from the beginning of the start
line and can extend past the end of that line, so the line/column computation
requires a bit of extra work.

The Rust tests were outdated, so the update does more than just add end-lines
and end-columns.